# api/swagger/swagger.yaml
definitions:
...
  github_com_deva-labs_univia_pkg_types.AssignPermissionRequest:
    properties:
      permission_ids:
        example:
        - '[]'
        items:
          type: string
        type: array
      role_id:
        example: 36byte
        type: string
    type: object
...
  github_com_deva-labs_univia_pkg_types.CreatePermissionRequest:
    properties:
      name:
        example: allow_create_post
        type: string
    type: object
...
  github_com_deva-labs_univia_pkg_types.LikeRequest:
    properties:
      post_id:
        type: string
    type: object
...
  github_com_deva-labs_univia_pkg_types.StatusBadRequest:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Bad Request
        type: string
    type: object
...
  github_com_deva-labs_univia_pkg_types.StatusInternalError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Internal Server Error
        type: string
    type: object
  github_com_deva-labs_univia_pkg_types.StatusOK:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Success
        type: string
    type: object
...
  github_com_deva-labs_univia_pkg_types.SuccessDisLikeAPostResponse:
    properties:
      data:
        properties:
          totalLikes:
            example: 1
            type: integer
        type: object
      status:
        $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusOK'
    type: object
  github_com_deva-labs_univia_pkg_types.SuccessLikeAPostResponse:
    properties:
      data:
        properties:
          totalLikes:
            example: 1
            type: integer
        type: object
      status:
        $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusOK'
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/likes:
    get:
      description: Like the post
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post ID Required
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.LikeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully Like A Post
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessLikeAPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      security:
      - BearerAuth: []
      summary: Like the post
      tags:
      - Social Routes
  /api/v1/likes/undo:
    get:
      description: DisLike the post
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Post ID Required
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.LikeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully Like A Post
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.SuccessDisLikeAPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusBadRequest'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.StatusInternalError'
      security:
      - BearerAuth: []
      summary: DisLike the post
      tags:
      - Social Routes
...
  /api/v1/permissions/assign:
    post:
      consumes:
      - application/json
      description: Grant multiple permissions to a specific role
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Role ID and list of Permission IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.AssignPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Assign permissions to a role
      tags:
      - Permissions
  /api/v1/permissions/create:
    post:
      consumes:
      - application/json
      description: Admin creates a new permission
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Permission name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_deva-labs_univia_pkg_types.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new permission
      tags:
      - Permissions
  /api/v1/permissions/list:
    post:
      description: Get all permissions in the system
      parameters:
      - description: Bearer <access_token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all permissions
      tags:
      - Permissions
...
swagger: "2.0"